Some missing polish related with evaluations v2 editor

+++ packages/core/src/services/evaluationsV2/llm/comparison.ts
@@ -96,7 +96,7 @@ async function validate(
 }
 
 const promptSchema = z.object({
-  score: z.number().min(0).max(100),
+  score: z.number().int().min(0).max(100),
   reason: z.string(),
 })
